pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-help"
version = "2024.10.1"
description = "A Django Knowledgebase"
authors = [{ name = "Jack Linke", email = "jacklinke@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Development Status :: 1 - Planning",
]
requires-python = ">=3.9,<4.0"
dependencies = [
    "django>=4.2",
    "click>=8.1.7",
    "django-markdownx>=4.0.7",
    "django-translated-fields>=0.13.0",
    "django-taggit>=6.1.0",
    "python-frontmatter>=1.1.0",
    "django-crispy-forms>=2.3",
]

[project.urls]
Homepage = "https://github.com/OmenApps/django-help"
Repository = "https://github.com/OmenApps/django-help"
Documentation = "https://django-help.readthedocs.io"
Changelog = "https://github.com/OmenApps/django-help/releases"

[tool.uv]
dev-dependencies = [
    "psycopg2-binary>=2.9.9",
    "Pygments>=2.18.0",
    "bandit>=1.7.8",
    "black>=24.4.2",
    "coverage[toml]>=7.5.1",
    "darglint>=1.8.1",
    "flake8==7.0.0",
    "flake8-bugbear>=24.4.26",
    "flake8-docstrings>=1.7.0",
    "flake8-rst-docstrings>=0.3.0",
    "furo>=2024.5.6",
    "isort>=5.13.2",
    "nox>=2024.4.15",
    "pep8-naming>=0.14.1",
    "pre-commit>=3.7.1",
    "pre-commit-hooks>=4.6.0",
    "pytest>=8.2.1",
    "pytest-cov>=5.0.0",
    "pytest-django>=4.8.0",
    "pyupgrade>=3.15.2",
    "safety>=3.2.0",
    "sphinx>=7.3.7",
    "sphinx-autobuild>=2024.4.16",
    "sphinx-click>=6.0.0",
    "xdoctest[colors]>=1.1.3",
    "myst-parser>=3.0.1",
]

[tool.black]
line-length = 120
target-version = ["py39", "py310", "py311", "py312"]
force-exclude = '''
(
  .nox
)
'''

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["example_project", "*/example_project"]

[tool.coverage.run]
branch = true
source = ["src", "example_project"]

[tool.coverage.report]
show_missing = true
fail_under = 50
omit = [".nox/*", "example_project/*", "**/migrations/*", "**/__init__.py"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "example_project.settings"
python_files = ["*test_*.py", "*_test.py", "example_project/*.py"]
log_cli = true
log_cli_level = "INFO"

[tool.isort]
profile = "black"
force_single_line = true
lines_after_imports = 2
extend_skip = [".nox"]
